Enhance service configuration and implement Elasticsearch refresh functionality
- Added ElasticsearchRefreshService to handle immediate visibility of changes in Elasticsearch for streams and notifications.
- Updated services.yaml to include the new ElasticsearchRefreshService with appropriate arguments.
- Commented out existing notification and option fixtures for clarity and future reference.
- Refactored various command handlers and services to utilize saveAndFlush for immediate persistence, ensuring data consistency across operations.

+++ src/Core/Application/Command/UpdateTaskSuccessCommand.php
@@ -12,7 +12,7 @@ final class UpdateTaskSuccessCommand extends CommandAbstract implements AsyncCom
 {
     public function __construct(
         private Uuid $taskId,
-        private float $processingTime,
+        private int $processingTime,
     ) {
     }
